refactor: rename storage stream module
